add BitVectorTest and fix a bug in rank table preparation

+++ src/main/java/org/utgenome/weaver/align/BWTransform.java
@@ -181,8 +181,8 @@ public class BWTransform implements Command
     public static void buildSuffixArray(IUPACSequence seq, File suffixArrayFile, File bwtFile) throws IOException {
 
         StopWatch timer = new StopWatch();
-        // TODO to make applicable more than 2GB sequence
-        int[] SA = new int[(int) seq.size()];
+        // TODO to make applicable more than 2GB sequencel
+        int[] SA = new int[seq.size()];
         {
             SAIS.suffixsort(seq, SA, 16);
             _logger.info("Sparse SA file: " + suffixArrayFile);
